+++ src/diffusers/models/unet_sde_score_estimation.py
@@ -27,7 +27,16 @@ from ..configuration_utils import ConfigMixin
 from ..modeling_utils import ModelMixin
 from .attention import AttentionBlock
 from .embeddings import GaussianFourierProjection, get_timestep_embedding
-from .resnet import Downsample2D, ResnetBlock2D, Upsample2D, downsample_2d, upfirdn2d, upsample_2d
+from .resnet import (
+    Downsample2D,
+    FirDownsample2D,
+    FirUpsample2D,
+    ResnetBlock2D,
+    Upsample2D,
+    downsample_2d,
+    upfirdn2d,
+    upsample_2d,
+)
 
 
 def _setup_kernel(k):
@@ -40,94 +49,74 @@ def _setup_kernel(k):
     return k
 
 
-def _upsample_conv_2d(x, w, k=None, factor=2, gain=1):
-    """Fused `upsample_2d()` followed by `Conv2d()`.
-
-    Args:
-    Padding is performed only once at the beginning, not between the operations. The fused op is considerably more
-    efficient than performing the same calculation using standard TensorFlow ops. It supports gradients of arbitrary
-    order.
-      x: Input tensor of the shape `[N, C, H, W]` or `[N, H, W,
-        C]`.
-      w: Weight tensor of the shape `[filterH, filterW, inChannels,
-        outChannels]`. Grouped convolution can be performed by `inChannels = x.shape[0] // numGroups`.
-      k: FIR filter of the shape `[firH, firW]` or `[firN]`
-        (separable). The default is `[1] * factor`, which corresponds to nearest-neighbor upsampling.
-      factor: Integer upsampling factor (default: 2). gain: Scaling factor for signal magnitude (default: 1.0).
-
-    Returns:
-      Tensor of the shape `[N, C, H * factor, W * factor]` or `[N, H * factor, W * factor, C]`, and same datatype as
-      `x`.
-    """
-
-    assert isinstance(factor, int) and factor >= 1
-
-    # Check weight shape.
-    assert len(w.shape) == 4
-    convH = w.shape[2]
-    convW = w.shape[3]
-    inC = w.shape[1]
-
-    assert convW == convH
-
-    # Setup filter kernel.
-    if k is None:
-        k = [1] * factor
-    k = _setup_kernel(k) * (gain * (factor**2))
-    p = (k.shape[0] - factor) - (convW - 1)
-
-    stride = (factor, factor)
-
-    # Determine data dimensions.
-    stride = [1, 1, factor, factor]
-    output_shape = ((x.shape[2] - 1) * factor + convH, (x.shape[3] - 1) * factor + convW)
-    output_padding = (
-        output_shape[0] - (x.shape[2] - 1) * stride[0] - convH,
-        output_shape[1] - (x.shape[3] - 1) * stride[1] - convW,
-    )
-    assert output_padding[0] >= 0 and output_padding[1] >= 0
-    num_groups = x.shape[1] // inC
-
-    # Transpose weights.
-    w = torch.reshape(w, (num_groups, -1, inC, convH, convW))
-    w = w[..., ::-1, ::-1].permute(0, 2, 1, 3, 4)
-    w = torch.reshape(w, (num_groups * inC, -1, convH, convW))
-
-    x = F.conv_transpose2d(x, w, stride=stride, output_padding=output_padding, padding=0)
-
-    return upfirdn2d(x, torch.tensor(k, device=x.device), pad=((p + 1) // 2 + factor - 1, p // 2 + 1))
+# def _upsample_conv_2d(x, w, k=None, factor=2, gain=1):
+#    """Fused `upsample_2d()` followed by `Conv2d()`. # # Args: # Padding is performed only once at the beginning, not between
+the operations. The fused op is considerably more # efficient than performing the same calculation using standard
+TensorFlow ops. It supports gradients of arbitrary # order. # x: Input tensor of the shape `[N, C, H, W]` or `[N, H, W,
+# C]`. # w: Weight tensor of the shape `[filterH, filterW, inChannels, # outChannels]`. Grouped convolution can be
+performed by `inChannels = x.shape[0] // numGroups`. # k: FIR filter of the shape `[firH, firW]` or `[firN]` #
+(separable). The default is `[1] * factor`, which corresponds to nearest-neighbor upsampling. # factor: Integer
+upsampling factor (default: 2). gain: Scaling factor for signal magnitude (default: 1.0). # # Returns: # Tensor of the
+shape `[N, C, H * factor, W * factor]` or `[N, H * factor, W * factor, C]`, and same datatype as # `x`. #"""
+#
+#    assert isinstance(factor, int) and factor >= 1
+#
+# Check weight shape.
+#    assert len(w.shape) == 4
+#    convH = w.shape[2]
+#    convW = w.shape[3]
+#    inC = w.shape[1]
+#
+#    assert convW == convH
+#
+# Setup filter kernel.
+#    if k is None:
+#        k = [1] * factor
+#    k = _setup_kernel(k) * (gain * (factor**2))
+#    p = (k.shape[0] - factor) - (convW - 1)
+#
+#    stride = (factor, factor)
+#
+# Determine data dimensions.
+#    stride = [1, 1, factor, factor]
+#    output_shape = ((x.shape[2] - 1) * factor + convH, (x.shape[3] - 1) * factor + convW)
+#    output_padding = (
+#        output_shape[0] - (x.shape[2] - 1) * stride[0] - convH,
+#        output_shape[1] - (x.shape[3] - 1) * stride[1] - convW,
+#    )
+#    assert output_padding[0] >= 0 and output_padding[1] >= 0
+#    num_groups = x.shape[1] // inC
+#
+# Transpose weights.
+#    w = torch.reshape(w, (num_groups, -1, inC, convH, convW))
+#    w = w[..., ::-1, ::-1].permute(0, 2, 1, 3, 4)
+#    w = torch.reshape(w, (num_groups * inC, -1, convH, convW))
+#
+#    x = F.conv_transpose2d(x, w, stride=stride, output_padding=output_padding, padding=0)
+#
+#    return upfirdn2d(x, torch.tensor(k, device=x.device), pad=((p + 1) // 2 + factor - 1, p // 2 + 1))
 
 
-def _conv_downsample_2d(x, w, k=None, factor=2, gain=1):
-    """Fused `Conv2d()` followed by `downsample_2d()`.
-
-    Args:
-    Padding is performed only once at the beginning, not between the operations. The fused op is considerably more
-    efficient than performing the same calculation using standard TensorFlow ops. It supports gradients of arbitrary
-    order.
-        x: Input tensor of the shape `[N, C, H, W]` or `[N, H, W,
-          C]`.
-        w: Weight tensor of the shape `[filterH, filterW, inChannels,
-          outChannels]`. Grouped convolution can be performed by `inChannels = x.shape[0] // numGroups`.
-        k: FIR filter of the shape `[firH, firW]` or `[firN]`
-          (separable). The default is `[1] * factor`, which corresponds to average pooling.
-        factor: Integer downsampling factor (default: 2). gain: Scaling factor for signal magnitude (default: 1.0).
-
-    Returns:
-        Tensor of the shape `[N, C, H // factor, W // factor]` or `[N, H // factor, W // factor, C]`, and same datatype
-        as `x`.
-    """
-
-    assert isinstance(factor, int) and factor >= 1
-    _outC, _inC, convH, convW = w.shape
-    assert convW == convH
-    if k is None:
-        k = [1] * factor
-    k = _setup_kernel(k) * gain
-    p = (k.shape[0] - factor) + (convW - 1)
-    s = [factor, factor]
-    x = upfirdn2d(x, torch.tensor(k, device=x.device), pad=((p + 1) // 2, p // 2))
-    return F.conv2d(x, w, stride=s, padding=0)
+# def _conv_downsample_2d(x, w, k=None, factor=2, gain=1):
+#    """Fused `Conv2d()` followed by `downsample_2d()`. # # Args: # Padding is performed only once at the beginning, not
+between the operations. The fused op is considerably more # efficient than performing the same calculation using
+standard TensorFlow ops. It supports gradients of arbitrary # order. # x: Input tensor of the shape `[N, C, H, W]` or
+`[N, H, W, # C]`. # w: Weight tensor of the shape `[filterH, filterW, inChannels, # outChannels]`. Grouped convolution
+can be performed by `inChannels = x.shape[0] // numGroups`. # k: FIR filter of the shape `[firH, firW]` or `[firN]` #
+(separable). The default is `[1] * factor`, which corresponds to average pooling. # factor: Integer downsampling factor
+(default: 2). gain: Scaling factor for signal magnitude (default: 1.0). # # Returns: # Tensor of the shape `[N, C, H //
+factor, W // factor]` or `[N, H // factor, W // factor, C]`, and same datatype # as `x`. #"""
+#
+#    assert isinstance(factor, int) and factor >= 1
+#    _outC, _inC, convH, convW = w.shape
+#    assert convW == convH
+#    if k is None:
+#        k = [1] * factor
+#    k = _setup_kernel(k) * gain
+#    p = (k.shape[0] - factor) + (convW - 1)
+#    s = [factor, factor]
+#    x = upfirdn2d(x, torch.tensor(k, device=x.device), pad=((p + 1) // 2, p // 2))
+#    return F.conv2d(x, w, stride=s, padding=0)
 
 
 def _variance_scaling(scale=1.0, in_axis=1, out_axis=0, dtype=torch.float32, device="cpu"):
@@ -181,46 +170,6 @@ class Combine(nn.Module):
             return h + y
         else:
             raise ValueError(f"Method {self.method} not recognized.")
-
-
-class FirUpsample2D(nn.Module):
-    def __init__(self, channels=None, out_channels=None, use_conv=False, fir_kernel=(1, 3, 3, 1)):
-        super().__init__()
-        out_channels = out_channels if out_channels else channels
-        if use_conv:
-            self.Conv2d_0 = Conv2d(channels, out_channels, kernel_size=3, stride=1, padding=1)
-        self.use_conv = use_conv
-        self.fir_kernel = fir_kernel
-        self.out_channels = out_channels
-
-    def forward(self, x):
-        if self.use_conv:
-            h = _upsample_conv_2d(x, self.Conv2d_0.weight, k=self.fir_kernel)
-            h = h + self.Conv2d_0.bias.reshape(1, -1, 1, 1)
-        else:
-            h = upsample_2d(x, self.fir_kernel, factor=2)
-
-        return h
-
-
-class FirDownsample2D(nn.Module):
-    def __init__(self, channels=None, out_channels=None, use_conv=False, fir_kernel=(1, 3, 3, 1)):
-        super().__init__()
-        out_channels = out_channels if out_channels else channels
-        if use_conv:
-            self.Conv2d_0 = self.Conv2d_0 = Conv2d(channels, out_channels, kernel_size=3, stride=1, padding=1)
-        self.fir_kernel = fir_kernel
-        self.use_conv = use_conv
-        self.out_channels = out_channels
-
-    def forward(self, x):
-        if self.use_conv:
-            x = _conv_downsample_2d(x, self.Conv2d_0.weight, k=self.fir_kernel)
-            x = x + self.Conv2d_0.bias.reshape(1, -1, 1, 1)
-        else:
-            x = downsample_2d(x, self.fir_kernel, factor=2)
-
-        return x
 
 
 class NCSNpp(ModelMixin, ConfigMixin):
